AbsoluteX/Y addressing modes now cast to u32 and back to u16 correctly

+++ src/cpu/cpu_executor.rs
@@ -68,11 +68,11 @@ impl CpuExecutor {
                 dr.value_final = Some(mem.read8(dr.addr_final.unwrap()).unwrap());
             },
             AddressMode::AbsoluteX       => {
-                dr.addr_final  = Some( (mem.read16(cpu_state.pc+1).unwrap() + cpu_state.x as u16) % 65535);
+                dr.addr_final  = Some( ((mem.read16(cpu_state.pc+1).unwrap() as u32 + cpu_state.x as u32) % 65535) as u16);
                 dr.value_final = Some(mem.read8(dr.addr_final.unwrap()).unwrap());
             },
             AddressMode::AbsoluteY       => {
-                dr.addr_final  = Some( (mem.read16(cpu_state.pc+1).unwrap() + cpu_state.y as u16) % 65535);
+                dr.addr_final  = Some( ((mem.read16(cpu_state.pc+1).unwrap() as u32 + cpu_state.y as u32) % 65535) as u16);
                 dr.value_final = Some(mem.read8(dr.addr_final.unwrap()).unwrap());
             },
             AddressMode::Immediate       => {
